requirements.txt
lxml
pyyaml
click
pytest
